se agrega dependencia redux-persist  se implementa combineReducer y se fixean los states de los componentes
