+++ client/package.json
@@ -20,6 +20,7 @@
     "react-redux": "^7.2.6",
     "react-router-dom": "^6.3.0",
     "react-scripts": "4.0.0",
+    "redux-persist": "^6.0.0",
     "redux-thunk": "^2.4.1",
     "sweetalert": "^2.1.2",
     "web-vitals": "^0.2.4"
