+++ client/src/components/Cart/Cart.jsx
@@ -1,6 +1,6 @@
 import s from "./cart.module.css";
 import { useReducer } from "react";
-import { cartInitialState, cartReducer } from "../../store/reducer-cart";
+import { cartInitialState, cartReducer } from "../../store/reducer/cartReducer";
 import { useDispatch } from "react-redux";
 import ProductItem from "../ProductItem/ProductItem";
 import { addToCart } from "../../store/actions-cart";
